Move all command files into bob.command pkg
