+++ src/main/java/bob/Parser.java
@@ -2,6 +2,17 @@ package bob;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
+import bob.command.Command;
+import bob.command.DeadlineCommand;
+import bob.command.DeleteCommand;
+import bob.command.EventCommand;
+import bob.command.ExitCommand;
+import bob.command.ListCommand;
+import bob.command.MarkCommand;
+import bob.command.TodoCommand;
+import bob.command.UnmarkCommand;
+import bob.command.WrongCommandException;
+
 public class Parser {
     public static Command parse(String userInput) throws WrongCommandException, IOException {
         if (userInput.equals("bye")) {
