+++ src/main/java/bob/command/Command.java
@@ -1,4 +1,7 @@
-package bob;
+package bob.command;
+
+import bob.Ui;
+
 public abstract class Command {
     public abstract void execute(Ui ui);
 }
