+++ src/main/java/bob/command/DeleteCommand.java
@@ -1,0 +1,30 @@
+package bob.command;
+import java.io.IOException;
+
+import bob.Task;
+import bob.TaskList;
+import bob.Ui;
+
+public class DeleteCommand extends Command {
+
+    private int number;
+
+    public DeleteCommand(int number) {
+        this.number = number;
+    }
+
+    @Override
+    public void execute(Ui ui) {
+        try {
+            Task task = TaskList.getTask(number);
+            TaskList.deleteTask(number);
+            ui.print("Bob is on it! Deleted this task: ", 
+                    task.toString(), 
+                    "Now you have " + TaskList.getCount() + " task(s).");
+        } catch (IndexOutOfBoundsException e) {
+            ui.print("Uh oh! Bob says...the task number does not exist.");
+        } catch (IOException e) {
+            ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
+        }   
+    }
+}
