+++ src/main/java/bob/command/ListCommand.java
@@ -1,4 +1,8 @@
-package bob;
+package bob.command;
+
+import bob.TaskList;
+import bob.Ui;
+
 public class ListCommand extends Command {
     @Override
     public void execute(Ui ui) {
