+++ src/main/java/bob/command/MarkCommand.java
@@ -1,19 +1,22 @@
-package bob;
+package bob.command;
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+import bob.TaskList;
+import bob.Ui;
+
+public class MarkCommand extends Command {
     
     private int number;
 
-    public UnmarkCommand(int number) {
+    public MarkCommand(int number) {
         this.number = number;
     }
     
     @Override
     public void execute(Ui ui) {
         try {
-            TaskList.unmarkTask(number);
-            ui.print("Bob is on it! Marked the following as undone [ ]", 
+            TaskList.markTask(number);
+            ui.print("Bob is on it! Marked the following as done [X]:", 
                     TaskList.getTask(number).toString());
         } catch (IndexOutOfBoundsException e) {
             ui.print("Uh oh! Bob says...the task number does not exist.");
