+++ src/main/java/bob/command/UnmarkCommand.java
@@ -1,26 +1,27 @@
-package bob;
+package bob.command;
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+import bob.TaskList;
+import bob.Ui;
 
+public class UnmarkCommand extends Command {
+    
     private int number;
 
-    public DeleteCommand(int number) {
+    public UnmarkCommand(int number) {
         this.number = number;
     }
-
+    
     @Override
     public void execute(Ui ui) {
         try {
-            Task task = TaskList.getTask(number);
-            TaskList.deleteTask(number);
-            ui.print("Bob is on it! Deleted this task: ", 
-                    task.toString(), 
-                    "Now you have " + TaskList.getCount() + " task(s).");
+            TaskList.unmarkTask(number);
+            ui.print("Bob is on it! Marked the following as undone [ ]", 
+                    TaskList.getTask(number).toString());
         } catch (IndexOutOfBoundsException e) {
             ui.print("Uh oh! Bob says...the task number does not exist.");
         } catch (IOException e) {
             ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
-        }   
+        }
     }
 }
