+++ src/main/java/bob/command/WrongCommandException.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.command;
 
 
 public class WrongCommandException extends Exception {
